[refactor] Move `GridItemCellViewModel` to single file
