+++ grid-cell/grid-cell/View/GridItemCell.swift
@@ -7,24 +7,6 @@
 
 import UIKit
 
-final class GridItemCellViewModel {
-  let title: String
-  let subtitle: String
-  let imageUrl: URL
-  let imageLoader: GridImageDataLoader
-  
-  init(item: GridItem, imageLoader: GridImageDataLoader) {
-    self.title = "\(item.id)"
-    self.subtitle = item.title
-    self.imageUrl = item.thumbnailUrl
-    self.imageLoader = imageLoader
-  }
-  
-  func loadImage(completion: @escaping (GridImageDataLoader.Result) -> Void) {
-    imageLoader.loadImageData(from: imageUrl, completion: completion)
-  }
-}
-
 final class GridItemCell: UICollectionViewCell {
   
   private lazy var baseImageView: UIImageView = {
